visitHistory minor refactoring
